// src/contacts/dtos/contacts.create.dto.ts
export class ContactsCreateDto {
  first_name: string;
  last_name: string;
  phone: string;
  email: string;
  password: string;
  address: string;
  type: string;
  ktp: string;
  ktpPhoto: string;
  peruri_type: string;
  npwp: string;
  npwpPhoto: string;
  selfPhoto: string;
  city: string;
  province: string;
  gender: string;
  placeofBirth: string;
  dateofBirth: string;
  orgUnit: string;
  workUnit: string;
  position: string;
}